@@ -1,55 +1,63 @@
 using System;
 using System.IO;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Extensions.Http;
-using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
+using System.Text.Json;
 using MrBigHead.Shared;
 using System.Collections.Generic;
+//using Microsoft.WindowsAzure.Storage.Table;
+using Microsoft.Azure.Functions.Worker;
+using Microsoft.Azure.Functions.Worker.Http;
+using Azure;
+using System.ComponentModel.DataAnnotations.Schema;
 using Microsoft.WindowsAzure.Storage.Table;
 
 namespace MrBigHead.Func
 {
-    public static class PostPhrases
+    public class PostPhrases
     {
-        [FunctionName("PostPhrases")]
-        public static async Task<IActionResult> Run(
-            [HttpTrigger(AuthorizationLevel.Function, "post", Route = null)] HttpRequest req,
-            [Table("Sayings")] CloudTable cloudTable,
-            ILogger log)
+        private readonly ILogger _logger;
+
+        public PostPhrases(ILoggerFactory loggerFactory)
         {
-            log.LogInformation("C# HTTP trigger function processed a request.");
+            _logger = loggerFactory.CreateLogger<PostPhrases>();
+        }
 
-            string requestBody = await new StreamReader(req.Body).ReadToEndAsync();
-            var data = JsonConvert.DeserializeObject<List<Saying>>(requestBody);
+        [Function("PostPhrases")]
+        public async Task<HttpResponseData> Run(
+            [HttpTrigger(AuthorizationLevel.Function, "post")] HttpRequestData req,
+            [TableInput("Sayings")] CloudTable cloudTable)
+        {
+            //_logger.LogInformation("C# HTTP trigger function processed a request.");
 
-            await InsertRecordsAsync(cloudTable, data);
+            //string requestBody = await new StreamReader(req.Body).ReadToEndAsync();
+            //var data = JsonSerializer.Deserialize<List<Saying>>(requestBody);
 
-            return new OkResult();
+            //await InsertRecordsAsync(cloudTable, data);
+
+            var response = req.CreateResponse(System.Net.HttpStatusCode.OK);
+            return response;
         }
 
         private static async Task InsertRecordsAsync(CloudTable cloudTable, List<Saying> data)
         {
-            foreach (var phrase in data)
-            {
-                var sayingEntity = new SayingEntity
-                {
-                    PartitionKey = phrase.Mood,
-                    Phrase = phrase.Phrase,
-                    RowKey = Guid.NewGuid().ToString()
-                };
+            //foreach (var phrase in data)
+            //{
+            //    var sayingEntity = new SayingEntity
+            //    {
+            //        PartitionKey = phrase.Mood,
+            //        Phrase = phrase.Phrase,
+            //        RowKey = Guid.NewGuid().ToString()
+            //    };
 
-                Console.WriteLine($"{sayingEntity.Phrase}");
+            //    Console.WriteLine($"{sayingEntity.Phrase}");
 
-                var tableOps = TableOperation.Insert(sayingEntity);
+            //    var tableOps = TableOperation.Insert(sayingEntity);
 
-                var tableResult = await cloudTable.ExecuteAsync(tableOps);
+            //    var tableResult = await cloudTable.ExecuteAsync(tableOps);
 
-                Console.WriteLine($"tableResult: {tableResult.HttpStatusCode}");
-            }
+            //    Console.WriteLine($"tableResult: {tableResult.HttpStatusCode}");
+            //}
         }
     }
 }
