@@ -1,18 +1,18 @@
-﻿using Microsoft.WindowsAzure.Storage;
+﻿using Azure;
+using Microsoft.WindowsAzure.Storage;
 using Microsoft.WindowsAzure.Storage.Table;
-using System.Collections.Generic;
-using System;
+using Adt = Azure.Data.Tables;
 
 namespace MrBigHead.Func
 {
-    public class SayingEntity : ITableEntity
+    public class SayingEntity : Adt.ITableEntity
     {
-        public string Mood { get; set; }
-        public string Phrase { get; set; }
+        public string? Mood { get; set; }
+        public string? Phrase { get; set; }
         public string PartitionKey { get => Mood; set => Mood = value; }
-        public string RowKey { get; set; }
-        public DateTimeOffset Timestamp { get; set; }
-        public string ETag { get; set; }
+        public string? RowKey { get; set; }
+        DateTimeOffset? Adt.ITableEntity.Timestamp { get; set; }
+        ETag Adt.ITableEntity.ETag { get; set; }
 
         public void ReadEntity(IDictionary<string, EntityProperty> properties, OperationContext operationContext)
         {
@@ -22,12 +22,6 @@ namespace MrBigHead.Func
         public IDictionary<string, EntityProperty> WriteEntity(OperationContext operationContext)
         {
             var results = new Dictionary<string, EntityProperty>();
-
-            //This don't work...  for now obvious reasons....
-            //results["PartitionKey"].StringValue = PartitionKey;
-            //results["Phrase"].StringValue = Phrase;
-            //results["RowKey"].StringValue = Guid.NewGuid().ToString();
-            //results["TimeStamp"].DateTime = DateTime.Now;
 
             results.Add("PartitionKey", EntityProperty.GeneratePropertyForString(PartitionKey));
             results.Add("Phrase", EntityProperty.GeneratePropertyForString(Phrase));
@@ -39,5 +33,4 @@ namespace MrBigHead.Func
             return results;
         }
     }
-
 }
